move PluginValidatorTask to public api
