+++ src/main/groovy/org/shipkit/internal/gradle/plugin/PluginValidationPlugin.java
@@ -1,11 +1,9 @@
 package org.shipkit.internal.gradle.plugin;
 
-
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.shipkit.internal.gradle.plugin.tasks.PluginValidatorTask;
-import org.shipkit.internal.gradle.util.JavaPluginUtil;
+import org.shipkit.gradle.plugin.PluginValidatorTask;
 import org.shipkit.internal.gradle.util.TaskMaker;
 
 /**
